streamlit_app/requirements.txt
pymilvus==2.2.8
msal_streamlit_authentication
azure-keyvault-secrets
azure-identity
extra-streamlit-components
microsoftgraph-python
psycopg2
pandas
pyGithub
asyncio
streamlit
openai
langchain==0.0.225
python-dotenv
tabulate
selenium
tiktoken
fuzzywuzzy